Increase max length for credentials from 64 chars
For machine-to-machine authentication it is better to use tokens than
harcoded passwords. In our stack we pass JWT tokens through the
WISTML password field instead of hardcoded passwords. The Witsml
Explorer UI caps the length of passwords to 64 characters, even if
servers typically excepts authorization headers up to 8k.

This commit changes the max password length to a bit less than the max
length of an authorization header. This way a JWT token can be used
as a password.